controllight_update: u8 bit for cmplw + clean stb/call (89.6->92.8)

+++ src/main/dll/LGT/dll_02AC_lgtcontrollight.c
@@ -79,7 +79,7 @@ void controllight_update(int obj)
 {
 
     ControlLightState* state = ((GameObject*)obj)->extra;
-    int bit = (u8)GameBit_Get(state->gameBit);
+    u8 bit = (u8)GameBit_Get(state->gameBit);
 
     if (bit != state->lastBit)
     {
